# admin/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Event Planner API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/tasks": {
      "get": {
        "tags": [
          "tasks",
          "tasks"
        ],
        "summary": "Get pending tasks for a messenger",
        "description": "Return pending tasks for a specific messenger.\n\nBots should poll this endpoint regularly (e.g. every minute) to discover\nnew work items.  Each task corresponds to a scheduled action (currently\nmailings) and is created separately for each messenger.  When a bot\ncompletes a task, it must call the completion endpoint with the task's\nID so that the task is not returned again.  Only administrators and\nbots with an administrator role can access tasks.\n\nParameters\n----------\nmessenger : str\n    Short code of the messenger requesting tasks (e.g. ``telegram``,\n    ``vk``, ``max``).  Tasks for other messengers are not returned.\nuntil : Optional[str]\n    ISO\u2011formatted timestamp limiting the maximum ``scheduled_at`` of\n    tasks returned.  Tasks scheduled after this time are excluded.\n    If omitted, the current server time is used.\n\nReturns\n-------\nList[TaskRead]\n    A list of pending tasks for the given messenger.",
        "operationId": "get_pending_tasks_api_v1_tasks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "messenger",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^(telegram|vk|max)$",
              "description": "Code of the messenger requesting tasks. Supported values: 'telegram', 'vk', 'max'. Only tasks for this messenger will be returned.",
              "title": "Messenger"
            },
            "description": "Code of the messenger requesting tasks. Supported values: 'telegram', 'vk', 'max'. Only tasks for this messenger will be returned."
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional ISO date/time up to which tasks should be returned. Tasks scheduled after this time are excluded. If omitted, the current server time is used.",
              "title": "Until"
            },
            "description": "Optional ISO date/time up to which tasks should be returned. Tasks scheduled after this time are excluded. If omitted, the current server time is used."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskRead"
                  },
                  "title": "Response Get Pending Tasks Api V1 Tasks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tasks/{task_id}/complete": {
      "post": {
        "tags": [
          "tasks",
          "tasks"
        ],
        "summary": "Mark a task as completed",
        "description": "Mark the specified task as completed.\n\nBots must call this endpoint after successfully processing a task returned\nby ``GET /tasks``.  Once a task is marked as completed it will no\nlonger be returned to any messenger, ensuring that each task is\nexecuted at most once per messenger.\n\nParameters\n----------\ntask_id : int\n    Identifier of the task to complete.  This ID is returned by the\n    ``GET /tasks`` endpoint.\n\nReturns\n-------\nNone\n    Responds with HTTP 204 No Content on success.",
        "operationId": "complete_task_api_v1_tasks__task_id__complete_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "scheduled_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduled At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type"
        ],
        "title": "TaskRead",
        "description": "Schema for reading a scheduled task.\n\nAttributes\n----------\nid : int\n    Unique identifier of the task in the corresponding table (e.g.,\n    ``mailings.id`` or ``support_tickets.id``).\ntype : str\n    A short code describing the type of task.  Current values are\n    ``\"mailing\"`` for scheduled mass messages and ``\"support_ticket\"``\n    for open support requests.  Clients should ignore types they\n    do not recognize.\ntitle : Optional[str]\n    A short human\u2011readable name for the task.  For mailings this\n    corresponds to the mailing title; for support tickets the\n    subject.\ndescription : Optional[str]\n    An optional longer description or content, e.g. the body of a\n    mailing.  Bots may choose to display this to operators or use\n    it as the basis of a message to users.\nscheduled_at : Optional[datetime]\n    For tasks that have a specific time when they should be\n    executed (e.g., mailings), this field contains the scheduled\n    date/time.  It may be ``None`` for tasks like open support\n    tickets."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}